+++ Changelog.md
@@ -8,17 +8,27 @@ and this project adheres to [Semantic Versioning].
 
 ## [2.0.7] - 2024-05-22
 
+### Added
+
+- World PVP and Open World PVP "queues" have been implemented for Active Queues.
+
 ### Changed
 
-- The current group leader's name will be shown instead of "YOUR LISTING".
+- [Active Queues] The current group leader's name will be shown instead of "YOUR LISTING".
 
 ### Fixed
 
 - [Last Invites] Restored the functionality to the "Last Invites" function and simplified it code-wise, making it easier to maintain in future.
 
+- [Entry Creation] Added the default backgrounds for all PVP categories and questing/custom.
+
 - [Search Panel & Application Viewer] Adjusted the sort button's position to match in both lite mode and normal mode.
 
 - [Application Viewer] Adjusted button sizes so the number of applicants is more visible.
+
+- [Application Viewer] The filters "Affix fit" and "Rating" are now also present in the application viewer.
+
+- [Filter Panel] Party, ress, lust and affix will now also be checked while loading the settings. 
 
 - [Active Queues] There shouldn't be anymore errors popping up being in a currently listed group and being infight.